Removes FreeEq File
